* Add memoization in r_print_zoom to speedup the work on big disks * Added R_NEW0 macro to fill with zeroes * Add documentation for [] keys in Visual mode

+++ libr/print/t/hex.c
@@ -8,7 +8,7 @@ int main()
 
 	r_cons_new();
 	p = r_print_new();
-	r_print_hexdump(p, (ut64)(main), buf, 128, 16, 1);
+	r_print_hexdump(p, (ut64)(size_t)(main), buf, 128, 16, 1);
 	r_cons_flush();
 
 	return 0;
